Move impl code from schema file into impl files
