+++ examples/CoffeeShop2/CoffeeShopSchema.cs
@@ -24,21 +24,10 @@ namespace CoffeeShop;
     E.g. changing 'productName' to 'name' can lead to the model being less stringent about what it places in that field. 
  */
 
-public class Cart
+public partial class Cart
 {
     [JsonPropertyName("items")]
     public CartItem[] Items { get; set; }
-
-    public virtual void GetUnknown(StringBuilder sb)
-    {
-        if (Items != null)
-        {
-            foreach (CartItem item in Items)
-            {
-                item.GetUnknown(sb);
-            }
-        }
-    }
 }
 
 [JsonPolymorphic]
@@ -47,22 +36,16 @@ public class Cart
 [JsonDerivedType(typeof(CoffeeDrink), typeDiscriminator: nameof(CoffeeDrink))]
 [JsonDerivedType(typeof(LatteDrink), typeDiscriminator: nameof(LatteDrink))]
 [JsonDerivedType(typeof(BakeryItem), typeDiscriminator: nameof(BakeryItem))]
-public abstract class CartItem
+public abstract partial class CartItem
 {
-    public virtual void GetUnknown(StringBuilder sb) { return; }
 }
 
 [Comment("Use this type for products with names that match NO listed PRODUCT NAME")]
-public class UnknownItem : CartItem
+public partial class UnknownItem : CartItem
 {
     [Comment("The text that wasn't understood")]
     [JsonPropertyName("text")]
     public string Text { get; set; }
-
-    public override void GetUnknown(StringBuilder sb)
-    {
-        sb.AppendLine(Text);
-    }
 }
 
 public abstract class LineItem : CartItem
@@ -71,7 +54,7 @@ public abstract class LineItem : CartItem
     public int Quantity { get; set; } = 1;
 }
 
-public class EspressoDrink : LineItem
+public partial class EspressoDrink : LineItem
 {
     [JsonVocab(Name = CoffeeShopVocabs.EspressoDrinks)]
     [JsonPropertyName("productName")]
@@ -86,11 +69,9 @@ public class EspressoDrink : LineItem
 
     [JsonPropertyName("options")]
     public DrinkOption[]? Options { get; set; }
-
-    public override void GetUnknown(StringBuilder sb) => Options.GetUnknown(sb);
 }
 
-public class CoffeeDrink : LineItem
+public partial class CoffeeDrink : LineItem
 {
     [JsonVocab(Name = CoffeeShopVocabs.CoffeeDrinks)]
     [JsonPropertyName("productName")]
@@ -105,11 +86,9 @@ public class CoffeeDrink : LineItem
 
     [JsonPropertyName("options")]
     public DrinkOption[]? Options { get; set; }
-
-    public override void GetUnknown(StringBuilder sb) => Options.GetUnknown(sb);
 }
 
-public class LatteDrink : LineItem
+public partial class LatteDrink : LineItem
 {
     [JsonVocab(Name = CoffeeShopVocabs.LatteDrinks)]
     [JsonPropertyName("productName")]
@@ -124,8 +103,6 @@ public class LatteDrink : LineItem
 
     [JsonPropertyName("options")]
     public DrinkOption[]? Options { get; set; }
-
-    public override void GetUnknown(StringBuilder sb) => Options.GetUnknown(sb);
 }
 
 public class BakeryItem : LineItem
